# src/internal/docs/swagger.yaml
basePath: /
definitions:
...
  models.BirdOutput:
    properties:
      description:
        type: string
      id:
        type: string
      image:
        $ref: '#/definitions/models.MediaOutput'
      name:
        type: string
      sound:
        $ref: '#/definitions/models.MediaOutput'
    type: object
  models.Err:
    properties:
      description:
        description: the error description for the developer, aka err.Error()
        type: string
      message:
        description: the displayed error text for the user
        type: string
      name:
        type: string
      status:
        type: integer
    type: object
  models.MediaOutput:
    properties:
      data:
        items:
          type: integer
        type: array
      fileType:
        type: string
      id:
        type: string
    type: object
  models.PostInput:
    properties:
      birdId:
        type: string
      createdAt:
        type: string
      imageId:
        type: string
      location:
        type: string
      soundId:
        type: string
      userId:
        type: string
    type: object
  models.PostOutput:
    properties:
      bird:
        $ref: '#/definitions/models.BirdOutput'
      createdAt:
        type: string
      id:
        type: string
      image:
        $ref: '#/definitions/models.MediaOutput'
      location:
        type: string
      sound:
        $ref: '#/definitions/models.MediaOutput'
      user:
        $ref: '#/definitions/models.UserOutput'
    type: object
  models.Response:
    properties:
      data: {}
      timestamp:
        type: string
    type: object
...
  models.UserOutput:
    properties:
      _id:
        type: string
      active:
        type: boolean
      createdAt:
        type: string
      username:
        type: string
    type: object
host: 127.0.0.1:3000
info:
  contact:
    email: ??
    name: Tovah Parnes
  description: A server for BirdAI API, for managing users, admin, birds, posts and
    more.
  license:
    name: MIT License
    url: https://opensource.org/license/mit/
  termsOfService: http://swagger.io/terms/
  title: BirdAI API
  version: "2.0"
paths:
...
  /posts/:
    post:
      consumes:
      - application/json
      description: Create a new post
      parameters:
      - description: post
        in: body
        name: set
        required: true
        schema:
          $ref: '#/definitions/models.PostInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      summary: Create a new post
      tags:
      - Posts
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete given post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      summary: Delete given post
      tags:
      - Posts
    get:
      consumes:
      - application/json
      description: Get post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.PostOutput'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "410":
          description: Gone
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      summary: Get post by ID
      tags:
      - Posts
    patch:
      consumes:
      - application/json
      description: Update given post
      parameters:
      - description: post ID
        in: path
        name: id
        required: true
        type: string
      - description: post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.PostInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      summary: Update given post
      tags:
      - Posts
  /posts/list:
    get:
      consumes:
      - application/json
      description: List all posts of a specified set
      parameters:
      - description: Set of posts
        in: query
        name: set
        type: integer
      - description: Search parameter for post
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.PostOutput'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      summary: List all posts of a specified set
      tags:
      - Posts
...
  /users/{id}/posts/list:
    get:
      consumes:
      - application/json
      description: List all posts of a specified set
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Set of posts
        in: query
        name: set
        type: integer
      - description: Search parameter for post
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.PostOutput'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      summary: List all posts of a specified set
      tags:
      - Posts
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: '>- Enter the token with the `Bearer: ` prefix, e.g. "Bearer eyJhbGciOiJIUzI1NiJ9.e30.DXCaKJxPJq8JhXOZIN4yWdBwJ4sraVbwH9imDCHZPaA".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"